Add files via upload
This is an assignment I completed in my Data Analysis for Finance class in the Fall of 2022. In this Jupyter Notebook, I define a function that parses through an SEC Quarterly Form data sheet and allows you to input any company's CIK. It will show you the company filed forms and the frequency of each document.