Rename UsingPandasOnTXTfile.ipynb to UsingPandasOnIDXfile.ipynb
In this assignment I used Pandas to extract data from the SEC website where I transformed the data to create functions that gave incites to which SEC forms a specific company submitted during the 3rd quarter of 2022. This function could provide actionable incites before investing in a publicly traded company.